xmlrpcpp
  * Changed source to SF 0.7 version

+++ src/xmlrpcpp/XmlRpcServer.cpp
@@ -1,28 +1,21 @@
 // ----------------------------------------------------------------------------
-// Copyright (C) 2014
-//              David Freese, W1HKJ
 //
-// This file is part of flrig.
-//
-// flrig is distributed in the hope that it will be useful,
-// but WITHOUT ANY WARRANTY; without even the implied warranty of
-// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.
-// ----------------------------------------------------------------------------
+// flxmlrpc Copyright (c) 2015 by W1HKJ, Dave Freese <iam_w1hkj@w1hkj.com>
+//    
 // XmlRpc++ Copyright (c) 2002-2008 by Chris Morley
-// This library is free software; you can redistribute it and/or
-// modify it under the terms of the GNU Lesser General Public
-// License as published by the Free Software Foundation; either
-// version 3 of the License, or (at your option) any later version.
-// 
-// This library is distributed in the hope that it will be useful,
-// but WITHOUT ANY WARRANTY; without even the implied warranty of
-// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU
-// Lesser General Public License for more details.
-// 
-// You should have received a copy of the GNU Lesser General Public License
-// along with this program.  If not, see <http://www.gnu.org/licenses/>.
 //
+// This file is part of fldigi
+//
+// flxmlrpc is free software; you can redistribute it and/or modify
+// it under the terms of the GNU Lesser General Public License as published by
+// the Free Software Foundation; either version 3 of the License, or
+// (at your option) any later version.
+//
+// You should have received a copy of the GNU General Public License
+// along with this program.  If not, see <http://www.gnu.org/licenses/>.
 // ----------------------------------------------------------------------------
+
+#include <config.h>
 
 #include "XmlRpcServer.h"
 #include "XmlRpcServerConnection.h"
@@ -501,7 +494,7 @@ XmlRpcServer::generateHeader(std::string const& body)
     "Content-length: ";
 
   char buffLen[40];
-  sprintf(buffLen,"%d\r\n\r\n", body.size());
+  sprintf(buffLen,"%d\r\n\r\n", static_cast<int>(body.size()));
 
   return header + buffLen;
 }
